requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
openai>=1.3.0
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.4.0
pytest-mock>=3.11.0
requests>=2.31.0
pydantic>=2.4.0